Created abstract animated cell to be reused in listViews to remove duplication of code
