+++ jvm/workbookapp/src/main/kotlin/org/wycliffeassociates/otter/jvm/workbookapp/ui/screens/ScriptureTakeCell.kt
@@ -18,14 +18,10 @@
  */
 package org.wycliffeassociates.otter.jvm.workbookapp.ui.screens
 
-import javafx.animation.TranslateTransition
-import javafx.event.EventHandler
-import javafx.scene.Node
-import javafx.scene.control.ListCell
-import javafx.util.Duration
 import org.wycliffeassociates.otter.common.data.workbook.Take
 import org.wycliffeassociates.otter.jvm.controls.card.EmptyCardCell
 import org.wycliffeassociates.otter.jvm.controls.card.ScriptureTakeCard
+import org.wycliffeassociates.otter.jvm.controls.card.events.AnimatedListCell
 import org.wycliffeassociates.otter.jvm.workbookapp.ui.model.TakeCardModel
 import tornadofx.*
 import java.text.MessageFormat
@@ -36,9 +32,9 @@ class ScriptureTakeCell(
     private val onDelete: (Take) -> Unit,
     private val onEdit: (Take) -> Unit,
     private val onSelected: (Take) -> Unit
-) : ListCell<TakeCardModel>() {
+) : AnimatedListCell<TakeCardModel>() {
 
-    private var takeCard = ScriptureTakeCard()
+    override val view = ScriptureTakeCard()
 
     override fun updateItem(item: TakeCardModel?, empty: Boolean) {
         super.updateItem(item, empty)
@@ -50,7 +46,7 @@ class ScriptureTakeCell(
             return
         }
 
-        graphic = takeCard.apply {
+        graphic = view.apply {
             takeProperty().set(item.take)
             audioPlayerProperty().set(item.audioPlayer)
             selectedProperty().set(item.selected)
@@ -77,55 +73,10 @@ class ScriptureTakeCell(
 
                 isAnimatingProperty.set(true)
                 animate(item) {
+                    view.isAnimatingProperty.set(false)
                     onSelected(item.take)
                 }
             }
         }
-    }
-
-    private fun animate(takeModel: TakeCardModel, callback: () -> Unit) {
-        shiftOtherNodes(takeModel)
-
-        val parentY = takeCard.parent.layoutY
-        takeCard.styleClass.add("selected")
-
-        // move selected node to top of the list
-        val ttUp = TranslateTransition(Duration.millis(600.0), takeCard)
-        ttUp.toY = -parentY
-        ttUp.onFinished = EventHandler {
-            takeCard.styleClass.remove("selected")
-            revertAnimation(takeCard) { takeCard.isAnimatingProperty.set(false) }
-            callback()
-        }
-        ttUp.play()
-    }
-
-    private fun shiftOtherNodes(takeModel: TakeCardModel) {
-        val selectedIndex = listView.items.indexOf(takeModel)
-        for (item in listView.items) {
-            if (listView.items.indexOf(item) < selectedIndex) {
-                moveDown(takeCard)
-            }
-        }
-    }
-
-    private fun moveDown(node: Node) {
-        val distance = node.boundsInLocal.height + 5
-        val tt = TranslateTransition(Duration.millis(600.0), node)
-        tt.byY = distance
-        tt.onFinished = EventHandler {
-            revertAnimation(node)
-        }
-        tt.play()
-    }
-
-    private fun revertAnimation(node: Node, onFinish: () -> Unit = { }) {
-        val distance = node.translateY
-        val ttRevertY = TranslateTransition(Duration.millis(1.0), node)
-        ttRevertY.byY = -distance
-        ttRevertY.onFinished = EventHandler {
-            onFinish()
-        }
-        ttRevertY.play()
     }
 }
